refactor: address review comments — extract mergePastSubjects, hoist disabledTabs arrays
